% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akde_barrier.R
\name{akde_barrier}
\alias{akde_barrier}
\title{Safe method for calculating an AKDE with a barrier}
\usage{
akde_barrier(
  data,
  CTMM,
  VMM = NULL,
  R = list(),
  barrier = NULL,
  barrier_in = FALSE,
  debias = TRUE,
  weights = FALSE,
  smooth = TRUE,
  error = 0.001,
  res = 10,
  grid = NULL,
  ...
)
}
\arguments{
\item{data}{2D timeseries telemetry data represented as a telemetry object or list of objects.}

\item{CTMM}{A ctmm movement model from the output of ctmm.fit or list of objects.}

\item{VMM}{An optional vertical ctmm object for 3D home-range calculation.}

\item{R}{A named list of raster covariates if CTMM contains an RSF model.}

\item{barrier}{An `sf` polygon data frame for enforcing hard boundaries}

\item{barrier_in}{Locations are assumed to be inside the `sf` polygons if `barrier_in=TRUE`
and outside of `barrier` if `barrier_in=FALSE`.}

\item{debias}{Debias the distribution for area estimation (AKDEc).}

\item{weights}{Optimally weight the data to account for sampling bias (See `\link[ctmm]{bandwidth}` for akde details).}

\item{smooth}{"Smooth" out errors from the data.}

\item{error}{Target probability error.}

\item{res}{Number of grid points along each axis, relative to the bandwidth.}

\item{grid}{Optional grid specification via raster, UD, or list of arguments (See ‘Details’ in `\link[ctmm]{akde}`).}

\item{...}{Arguments passed to akde, `\link[ctmm]{bandwidth}`, and `\link[ctmm]{mean.ctmm}`.}
}
\description{
Safe method for calculating an AKDE with a barrier
}
