% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmmUtils-package.R
\docType{package}
\name{crawlUtils-package}
\alias{ctmmUtils}
\alias{ctmmUtils-package}
\alias{crawlUtils-package}
\alias{crawlUtils}
\title{Functions To Increase Usability Of The \code{ctmm} Package}
\description{
This package is a collection of functions that enhance the \code{ctmm} package for
for analysis of animal telemetry data.

\tabular{ll}{
Package: \tab crawlUtils\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9007\cr
Date: \tab April 11, 2024\cr
License: \tab CC0 \cr
LazyLoad: \tab yes\cr
}
}
\note{
This software package is developed and maintained by scientists at the
NOAA Fisheries Pacific Islands Fisheries Science Center and should be
considered a fundamental research communication. The recommendations and
conclusions presented here are those of the authors and this software should
not be construed as official communication by NMFS, NOAA, or the U.S. Dept.
of Commerce. In addition, reference to trade names does not imply endorsement
by the National Marine Fisheries Service, NOAA. While the best efforts have
been made to insure the highest quality, tools such as this are under
constant development and are subject to change.
}
\author{
Devin S. Johnson and Josh M. London
Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
