% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_det.R
\name{migration_det}
\alias{migration_det}
\title{Migration detection}
\usage{
migration_det(
  data,
  min_disp,
  max_num_mig = 1,
  min_phase_len = 3,
  grid_res = "day",
  base = "first",
  max_k = 100
)
}
\arguments{
\item{data}{A `telemytry` data object of locations (see `\link[ctmm]{as.telemetry}`)}

\item{min_disp}{The minimum dispersion rate to be considered a migration interval,
e.g. 10 for a 10km dispersion minimum.}

\item{max_num_mig}{The maximum number of migration intervals.}

\item{min_phase_len}{The minimum length of time that a migration or non-migration event
will take, e.g., 7 implies a minimum of 7 time intervals for a phase.}

\item{grid_res}{The temporal resolution at which migrations are detected. e.g., "day" (default) implies migration
start and end is detected on a daily resolution.}

\item{base}{The location at which dispersion is measured. Can be one of \code{"first"} (first location),
\code{"last"} (final location), or some other \code{sf::sfc} point location.}

\item{max_k}{The maximum degrees of freedom used by \code{mgcv::gam} to model dispersion and estimate
the derivative of the dispersion function.}
}
\description{
Creates a data table that indicates the times of different
phases of movement. This method uses changes in the overall dispersion rate of the animal
from the 'base' time to detect changes in overall movement from small scale local movement
to large scale migration.
}
\author{
Devin S. Johnson
}
