% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouf_covmat.R
\name{ouf_corfun}
\alias{ouf_corfun}
\title{Create covariance function for a fitted OU or OUF ctmm model object}
\usage{
ouf_corfun(x)
}
\arguments{
\item{x}{A `ctmm` object created by a call to \code{\link[ctmm]{ctmm.fit}} or \code{\link[ctmm]{ctmm.select}}}
}
\description{
A function is created to evaluate the covariance function of the fitted
OUF movement model
}
\details{
The function returns a function to evaluate the covariance of the fitted
Ornstein-Ulenbeck Foraging movement model. The returned function has 2 arguments: (1) \code{s}
and (2) \code{t} both vectors of times to evaluate the covariance function of the fitted
OUF model.
}
\author{
Devin S. Johnson
}
