% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouf_covmat.R
\name{ouf_covmat}
\alias{ouf_covmat}
\title{Calculate correlation matrix for a set of times from a `ctmm` covariance function}
\usage{
ouf_covmat(x, times, inverse = FALSE)
}
\arguments{
\item{x}{A \code{ctmm} model object from a call to \code{\link[ctmm]{ctmm.fit}}
or `\link[ctmm]{ctmm.select}`.}

\item{times}{A vector of POSIX times at which the covariance matrix will be constructed.
A `telemetry` data object will also work.}

\item{inverse}{Logical. Should the inverse covariance matrix be returned. Defaults to `FALSE`}
}
\description{
Using a correlation function created by \code{\link{ouf_corfun}}
from a fitted OUF related model a covariance (correlation) matrix is created for observations
at the user provided times.
}
\author{
Devin S. Johnson
}
