% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouf_ess.R
\name{ouf_ess}
\alias{ouf_ess}
\title{Calculate Effective Sample Size for a Set of OUF locations}
\usage{
ouf_ess(x, times)
}
\arguments{
\item{x}{A \code{ctmm} object (See \code{\link[ctmm]{ctmm.fit}}).}

\item{times}{A vector of POSIX times. A `telemtry` data object will also work.}
}
\description{
Estimates the number of independent locations in a `ctmm` data set
using the mutual information method of Bartoszek (2016).
}
\details{
This function uses the "mutual information" effective sample size of Bartoszek (2016) to calculate the
equivalent number of independent animal locations. It also calculates individual contributions of each location using the
regression effective sample size in Bartoszek (2016). The output is a named list with `Ne` equal to the overall sample
and `w` is a vector of weights that sum to 1 overall. If you want the ESS value of each observation `Ne * w` will provide it.
}
\references{
Bartoszek, K. (2016). Phylogenetic effective sample size.
Journal of Theoretical Biology. 407:371-386. (See https://arxiv.org/pdf/1507.07113.pdf).
}
\author{
Devin S. Johnson
}
