% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouf_parallel_funs.R
\name{ouf_ess_parallel}
\alias{ouf_ess_parallel}
\title{Parallel ESS calculation of ctmm OUF family models}
\usage{
ouf_ess_parallel(fits, tdata, ...)
}
\arguments{
\item{fits}{A list of fitted `ctmm` OUF family models}

\item{tdata}{A list of telemetry data objects.}

\item{...}{Additional arguments (besides `data` and `CTMM`!) passed to `\link[ctmm]{ctmm.select}`}
}
\description{
Parallel ESS calculation of ctmm OUF family models
}
