% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_disp.R
\name{plot_disp}
\alias{plot_disp}
\title{Plot migration detection results}
\usage{
plot_disp(data, migr_tbl, interactive = FALSE)
}
\arguments{
\item{data}{Original data used by `\link[ctmmUtils]{migration_det}` call.}

\item{migr_tbl}{Results table produced by `\link[ctmmUtils]{migration_det}`.}

\item{interactive}{Logical. If `TRUE` the `plotly` package will be used to make the plot more interactive.}
}
\description{
PLot the animal dispersion from the base location over time. Points
are colored to reflect estimated migration and non-migration phases.
See \code{\link{migration_det}}
}
\author{
Devin S. Johnson
}
