% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telem_sda.R
\name{telem_sda}
\alias{telem_sda}
\title{Filter track for speed, distance and angle.}
\usage{
telem_sda(data, ..., filter = TRUE)
}
\arguments{
\item{data}{`telemetry` data object from the `ctmm` package.}

\item{...}{Additional arguments for the `\link[trip]{sda}` function. The argument `smax` (in km/hr) is
necessary. Please see the `\link[trip]{sda}` documentation.}

\item{filter}{Logical. Remove locations determined to be outliers based on the SDA filter. If `filter=FALSE` then a
column named `keep` will indicate which observations should be kept, but all will remain in the data. The default is
`filter=TRUE`}
}
\description{
Applies the Freitas SDA (speed, distance, angle) filter to assess telemetry data outliers. The
`\link[trip]{sda}` function from the `trip` package is used for computation.
}
\references{
Freitas, C., Lydersen, C., Fedak, M. A. and Kovacs, K. M. (2008),
A simple new algorithm to filter marine mammal Argos locations.
Marine Mammal Science, 24: 315?V325. doi: 10.1111/j.1748-7692.2007.00180.x
}
\author{
Devin S. Johnson
}
