% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_area_table.R
\name{ud_area_table}
\alias{ud_area_table}
\title{Get UD Area Table}
\usage{
ud_area_table(uds, level.UD = 0.95, unit = NULL)
}
\arguments{
\item{uds}{A named list of `ctmm` `UD` objects}

\item{level.UD}{A vector of UD levels for computation}

\item{unit}{Units for area estimates. Defaults to `unit = "m^2"`}
}
\description{
Get UD Area Table
}
