% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouf_parallel_funs.R
\name{ctmm_select_parallel}
\alias{ctmm_select_parallel}
\title{Parallel fitting of ctmm OUF family models}
\usage{
ctmm_select_parallel(tdata, ...)
}
\arguments{
\item{tdata}{A list of telemetry data objects.}

\item{...}{Additional arguments (besides `data` and `CTMM`!) passed to `\link[ctmm]{ctmm.select}`}
}
\description{
Parallel fitting of ctmm OUF family models
}
